+++ Kernel/GenericInterface/Operation/Ticket/TicketSearch.pm
@@ -133,16 +133,29 @@ perform TicketSearch Operation. This will return a Ticket ID list.
         #   At least one operator must be specified. Operators will be connected with AND,
         #       values in an operator with OR.
         #   You can also pass more than one argument to an operator: ['value1', 'value2']
-        DynamicField_FieldNameX => {
-            Empty             => 1,                       # will return dynamic fields without a value
-                                                          # set to 0 to search fields with a value present
-            Equals            => 123,
-            Like              => 'value*',                # "equals" operator with wildcard support
-            GreaterThan       => '2001-01-01 01:01:01',
-            GreaterThanEquals => '2001-01-01 01:01:01',
-            SmallerThan       => '2002-02-02 02:02:02',
-            SmallerThanEquals => '2002-02-02 02:02:02',
-        }
+        DynamicField => [                                                  # optional
+            {
+                Name   => 'some name',
+                Empty             => 1,                       # will return dynamic fields without a value
+                                                                  # set to 0 to search fields with a value present
+                Equals            => 123,
+                Like              => 'value*',                # "equals" operator with wildcard support
+                GreaterThan       => '2001-01-01 01:01:01',
+                GreaterThanEquals => '2001-01-01 01:01:01',
+                SmallerThan       => '2002-02-02 02:02:02',
+                SmallerThanEquals => '2002-02-02 02:02:02',
+            },
+            # ...
+        ],
+        # or
+        # DynamicField => {
+        #    Name   => 'some name',
+        #    # ...
+        #    Equals => 123,
+        #    # ...
+        #},
+
+
 
         # article stuff (optional)
         From    => '%spam@example.com%',
@@ -523,26 +536,47 @@ sub _GetDynamicFields {
         ObjectType => ['Ticket'],
     );
 
-    for my $ParameterName ( sort keys %Param ) {
-        if ( $ParameterName =~ m{\A DynamicField_ ( [a-zA-Z\d]+ ) \z}xms ) {
+    my %DynamicFieldsRaw;
+    if ( $Param{DynamicField} ) {
+        my %SearchParams;
+        if ( IsHashRefWithData( $Param{DynamicField} ) ) {
+            $DynamicFieldsRaw{ $Param{DynamicField}->{Name} } = $Param{DynamicField};
+        }
+        elsif ( IsArrayRefWithData( $Param{DynamicField} ) ) {
+            %DynamicFieldsRaw = map { $_->{Name} => $_ } @{ $Param{DynamicField} };
+        }
+        else {
+            return %DynamicFieldSearchParameters;
+        }
 
-            # loop over the dynamic fields configured
-            DYNAMICFIELD:
-            for my $DynamicFieldConfig ( @{ $Self->{DynamicField} } ) {
-                next DYNAMICFIELD if !IsHashRefWithData($DynamicFieldConfig);
-                next DYNAMICFIELD if !$DynamicFieldConfig->{Name};
+    }
+    else {
 
-                # skip all fields that does not match with current field name ($1)
-                # without the 'DynamicField_' prefix
-                next DYNAMICFIELD if $DynamicFieldConfig->{Name} ne $1;
-
-                # set search parameter
-                $DynamicFieldSearchParameters{ 'DynamicField_' . $DynamicFieldConfig->{Name} }
-                    = $Param{ 'DynamicField_' . $DynamicFieldConfig->{Name} };
-
-                last DYNAMICFIELD;
+        # Compatibility with older versions of the web service.
+        for my $ParameterName ( sort keys %Param ) {
+            if ( $ParameterName =~ m{\A DynamicField_ ( [a-zA-Z\d]+ ) \z}xms ) {
+                $DynamicFieldsRaw{$1} = $Param{$ParameterName};
             }
         }
+    }
+
+    # loop over the dynamic fields configured
+    DYNAMICFIELD:
+    for my $DynamicFieldConfig ( @{ $Self->{DynamicField} } ) {
+        next DYNAMICFIELD if !IsHashRefWithData($DynamicFieldConfig);
+        next DYNAMICFIELD if !$DynamicFieldConfig->{Name};
+
+        # skip all fields that does not match with current field name
+        next DYNAMICFIELD if !$DynamicFieldsRaw{ $DynamicFieldConfig->{Name} };
+
+        next DYNAMICFIELD if !IsHashRefWithData( $DynamicFieldsRaw{ $DynamicFieldConfig->{Name} } );
+
+        my %SearchOperators = %{ $DynamicFieldsRaw{ $DynamicFieldConfig->{Name} } };
+
+        delete $SearchOperators{Name};
+
+        # set search parameter
+        $DynamicFieldSearchParameters{ 'DynamicField_' . $DynamicFieldConfig->{Name} } = \%SearchOperators;
     }
 
     # allow free fields